email verification logic remove in repository file and add into business file
